+++ src/pages/hyunminChoi/Main/Main.js
@@ -1,14 +1,7 @@
 import React, { useState, useEffect } from 'react';
-import {
-  FaComment,
-  FaEllipsisV,
-  FaHeart,
-  FaPaperPlane,
-  FaStickyNote,
-} from 'react-icons/fa';
 import Nav from '../../../../src/components/Nav/Nav.js';
 import MainRightStory from '../components/MainRightStory.js';
-import CommentList from './CommentList.js';
+import Feed from '../components/Feed.js';
 import './Main.scss';
 
 function Main() {
@@ -30,23 +23,13 @@ function Main() {
     },
   ];
 
-  let [userName] = useState('hacker');
-  let [comment, setComment] = useState('');
-  let [feedComments, setFeedComments] = useState([]);
-  let [isValid, setIsValid] = useState(false);
-
-  let post = e => {
-    const copyFeedComments = [...feedComments];
-    copyFeedComments.push(comment);
-    setFeedComments(copyFeedComments);
-    setComment('');
-  };
+  const [feed, setFeed] = useState([]);
 
   useEffect(() => {
-    fetch('data/commentData.json', {})
+    fetch('data/feedData.json', {})
       .then(res => res.json())
       .then(data => {
-        setFeedComments(data);
+        setFeed(data);
       });
   }, []);
 
@@ -57,98 +40,15 @@ function Main() {
         <div className="wrapper">
           <div className="main">
             {/* 피드영역 */}
-            <div className="feeds">
-              <article>
-                <div className="content">
-                  <div className="top">
-                    <div className="topProfile">
-                      <img
-                        src="../../../../images/hyunminChoi/Main/developer.avif"
-                        alt="profile"
-                        className="profileImg"
-                      />
-                      <div>influencer</div>
-                    </div>
-                    <div className="topService">
-                      <FaEllipsisV />
-                    </div>
-                  </div>
+            {feed.map(feedList => (
+              <Feed
+                key={feedList.id}
+                profileName={feedList.userName}
+                userPhoto={feedList.userPhoto}
+                feedPhoto={feedList.feedPhoto}
+              />
+            ))}
 
-                  <div className="mid">
-                    <img
-                      src="../../../../images/hyunminChoi/Main/couple.avif"
-                      alt="커플 사진"
-                    />
-                  </div>
-
-                  <div className="bottom">
-                    <section className="convenience">
-                      <span>
-                        <FaHeart />
-                      </span>
-                      <span>
-                        <FaComment />
-                      </span>
-                      <span>
-                        <FaPaperPlane />
-                      </span>
-                      <span>
-                        <FaStickyNote />
-                      </span>
-                    </section>
-                    <section className="like">
-                      <p>좋아요 1,092,182개</p>
-                    </section>
-                    <div className="commentBox">
-                      {feedComments.map(commentArr => {
-                        return (
-                          <CommentList
-                            userName={userName}
-                            userComment={commentArr}
-                            feedComments={feedComments}
-                            setFeedComments={setFeedComments}
-                          />
-                        );
-                      })}
-                    </div>
-                    <section className="dateAgo">3일전</section>
-                    <section className="addCommentBox">
-                      <input
-                        type="text"
-                        className="inputComment"
-                        placeholder="댓글 달기..."
-                        onChange={e => {
-                          setComment(e.target.value);
-                        }}
-                        onKeyUp={e => {
-                          e.target.value.length > 0
-                            ? setIsValid(true)
-                            : setIsValid(false);
-                        }}
-                        onKeyDown={e => {
-                          if (e.key === 'Enter' && isValid) {
-                            post();
-                          }
-                        }}
-                        value={comment}
-                      />
-                      <button
-                        type="button"
-                        className={
-                          comment.length > 0
-                            ? 'submitCommentActive'
-                            : 'submitCommentInactive'
-                        }
-                        onClick={post}
-                        disabled={isValid ? false : true}
-                      >
-                        게시
-                      </button>
-                    </section>
-                  </div>
-                </div>
-              </article>
-            </div>
             {/* 우측메뉴 */}
             <div className="mainRight">
               <div className="mainRight_account">
