+++ src/pages/hyunminChoi/Main/Main.scss
@@ -23,6 +23,7 @@
 
   .main {
     display: flex;
+    flex-direction: column;
     width: 70%;
 
     /* feed 영역 */
